# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad Seller System API",
    "description": "IAB OpenDirect 2.1 compliant seller agent for programmatic advertising. Supports product discovery, tiered pricing, proposal evaluation, multi-round negotiation, deal execution, order management, and change requests.",
    "contact": {
      "name": "IAB Tech Lab",
      "url": "https://iabtechlab.com/"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/proposals": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Submit Proposal",
        "description": "Submit a proposal for review.",
        "operationId": "submit_proposal_proposals_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}/counter": {
      "post": {
        "tags": [
          "Negotiation"
        ],
        "summary": "Counter Proposal",
        "description": "Submit a counter-offer in an ongoing negotiation.\n\nLoads or creates a NegotiationHistory, evaluates the buyer's offer,\npersists the updated history, and emits a NEGOTIATION_ROUND event.\n\nEP-5.2: this endpoint previously applied NO trust-tier ceiling \u2014 a\nbuyer could self-assert ADVERTISER pricing by populating\nadvertiser_id. The claimed tier is now verified against the agent\nregistry (agent_url) or the API key; unverifiable claims floor.",
        "operationId": "counter_proposal_proposals__proposal_id__counter_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterOfferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}/negotiation": {
      "get": {
        "tags": [
          "Negotiation"
        ],
        "summary": "Get Negotiation Status",
        "description": "Get full negotiation history for a proposal.",
        "operationId": "get_negotiation_status_proposals__proposal_id__negotiation_get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/negotiations/messages": {
      "post": {
        "tags": [
          "Negotiation"
        ],
        "summary": "Post Negotiation Message",
        "description": "Canonical negotiation surface \u2014 accepts the shared NegotiationMessage.\n\nThe seller now validates the SAME message the buyer emits: a required\n``action`` enum (accept/counter/reject/final_offer) plus ``buyer_price``\nas :class:`Money`. A well-formed counter no longer 422s (the historical\nbug). Priced actions (counter/final_offer) run the seller's untouched\nnegotiation engine; accept/reject are recorded as terminal rounds off\nthe existing history.",
        "operationId": "post_negotiation_message_api_v1_negotiations_messages_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationRoundResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quotes": {
      "post": {
        "tags": [
          "Quotes"
        ],
        "summary": "Create Quote",
        "description": "Request a non-binding price quote from the seller.\n\nThe seller evaluates the request against existing pricing rules and\nreturns a quote with pricing, terms, and availability. Quotes are\nephemeral with a 24-hour TTL \u2014 no Deal ID is created.",
        "operationId": "create_quote_api_v1_quotes_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{quote_id}": {
      "get": {
        "tags": [
          "Quotes"
        ],
        "summary": "Get Quote",
        "description": "Retrieve a previously issued quote.\n\nReturns 410 Gone if the quote has expired.",
        "operationId": "get_quote_api_v1_quotes__quote_id__get",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CounterOfferRequest": {
        "properties": {
          "buyer_price": {
            "type": "number",
            "title": "Buyer Price"
          },
          "buyer_tier": {
            "type": "string",
            "title": "Buyer Tier",
            "default": "public"
          },
          "agency_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Id"
          },
          "advertiser_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertiser Id"
          },
          "agent_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Url"
          }
        },
        "type": "object",
        "required": [
          "buyer_price"
        ],
        "title": "CounterOfferRequest",
        "description": "Request to submit a counter-offer in a negotiation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Money": {
        "properties": {
          "amount_micros": {
            "type": "integer",
            "title": "Amount Micros",
            "description": "Amount in micros; 1,000,000 micros = 1 currency unit."
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "description": "ISO 4217 alpha-3 currency code.",
            "default": "USD"
          }
        },
        "type": "object",
        "required": [
          "amount_micros"
        ],
        "title": "Money",
        "description": "Exact money amount in integer micros (flagged decision FD-11).\n\n``1_000_000`` micros = 1 currency unit \u2014 the ad-industry convention\n(Google Ad Manager, among others, prices in micros). Float is BANNED on\nthe wire for money: IEEE 754 floating point is non-deterministic for\nmoney math (``0.1 + 0.2 != 0.3``, and repeated CPM \u2014 cost per mille \u2014\narithmetic accumulates error), and the two source repos used ``float``\nend-to-end; that defect must not be fossilized into the spec. Every\nprice, rate, budget, and offer in the shared contract is a ``Money``.\n\n``amount_micros`` is a strict integer: float inputs are rejected at\nvalidation time rather than silently truncated."
      },
      "NegotiationAction": {
        "type": "string",
        "enum": [
          "accept",
          "counter",
          "reject",
          "final_offer"
        ],
        "title": "NegotiationAction",
        "description": "Action taken in a negotiation round."
      },
      "NegotiationMessage": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key",
            "description": "Requester-minted opaque key (UUID recommended). Same key -> same response, no duplicate side effects (FD-12). Reusing a key with a different body is an idempotency_conflict error."
          },
          "action": {
            "$ref": "#/components/schemas/NegotiationAction",
            "description": "REQUIRED move discriminator: 'accept', 'counter', 'reject' (walk-away), or 'final_offer'. No default \u2014 the action-less bare-price payload that caused the historical 422 does not validate."
          },
          "negotiation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negotiation Id",
            "description": "Seller-issued id of the negotiation to continue; None opens a new negotiation on proposal_id/quote_id."
          },
          "proposal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id",
            "description": "Proposal under negotiation, if proposal-led."
          },
          "quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Id",
            "description": "Quote under negotiation, if quote-led."
          },
          "round_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Round Number",
            "description": "The round the sender believes it is answering, for optimistic concurrency; a mismatch is a 'contention' error. The seller's numbering is authoritative."
          },
          "buyer_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ],
            "description": "The buyer's price this round (exact micros; FD-11). REQUIRED for 'counter'/'final_offer'; optional echo on 'accept'; omitted on 'reject'."
          },
          "buyer_identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuyerIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "description": "Optional human-readable rationale.",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "idempotency_key",
          "action"
        ],
        "title": "NegotiationMessage",
        "description": "A buyer negotiation move (money-mutating: FD-12).\n\nExactly one negotiation context is required: ``negotiation_id`` to\ncontinue, or ``proposal_id``/``quote_id`` to open (the seller mints\n``negotiation_id`` on open)."
      },
...
      "NegotiationRoundResponse": {
        "properties": {
          "negotiation_id": {
            "type": "string",
            "title": "Negotiation Id",
            "description": "Seller-issued negotiation identifier."
          },
          "status": {
            "$ref": "#/components/schemas/NegotiationStatus",
            "description": "'active' or terminal ('accepted'/'rejected'/'expired'); terminal negotiations refuse further messages (negotiation_closed)."
          },
          "round": {
            "$ref": "#/components/schemas/NegotiationRound"
          },
          "rounds_remaining": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Remaining",
            "description": "Rounds left before the seller walks away, if disclosed."
          }
        },
        "type": "object",
        "required": [
          "negotiation_id",
          "status",
          "round"
        ],
        "title": "NegotiationRoundResponse",
        "description": "The seller's answer to a :class:`NegotiationMessage`.\n\nEmbeds the shared\n:class:`~iab_agentic_primitives.primitives.NegotiationRound` primitive\n\u2014 the same record appended to the Negotiation history \u2014 so the two\nsides cannot disagree about what a round contains. ``round.action`` is\nthe SELLER's move (accept/counter/reject/final_offer); when\n``status`` is terminal the round is the last one."
      },
      "NegotiationStatus": {
        "type": "string",
        "enum": [
          "active",
          "accepted",
          "rejected",
          "expired"
        ],
        "title": "NegotiationStatus",
        "description": "Status of a negotiation (typed; was a raw string in the seller repo)."
      },
...
      "ProposalRequest": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "deal_type": {
            "type": "string",
            "title": "Deal Type"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "impressions": {
            "type": "integer",
            "title": "Impressions"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "buyer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Id"
          },
          "agency_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Id"
          },
          "advertiser_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertiser Id"
          },
          "agent_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Url"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "deal_type",
          "price",
          "impressions",
          "start_date",
          "end_date"
        ],
        "title": "ProposalRequest",
        "description": "Request to submit a proposal."
      },
      "ProposalResponse": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "counter_terms": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Terms"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "pricing_verified": {
            "type": "boolean",
            "title": "Pricing Verified",
            "default": false
          },
          "pricing_verification_reason": {
            "type": "string",
            "title": "Pricing Verification Reason",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "proposal_id",
          "recommendation",
          "status"
        ],
        "title": "ProposalResponse",
        "description": "Proposal submission response."
      },
...
      "QuoteResponse": {
        "properties": {
          "quote": {
            "$ref": "#/components/schemas/Quote"
          }
        },
        "type": "object",
        "required": [
          "quote"
        ],
        "title": "QuoteResponse",
        "description": "Success envelope for the quote endpoints: wraps the Quote primitive.\n\nThe quote carries its own ``media_type``, ``linear_tv`` details,\n``rate_card_id``, pricing, terms, availability, and ``expires_at``\n(quotes are ephemeral \u2014 the seller enforces a TTL and answers a\n``quote_expired`` error after it elapses)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}